Undoing some changes
